requirements.txt


# Backend
fastapi
uvicorn[standard]
pydantic
python-dotenv
pymongo
pdfplumber
numpy
tiktoken
google-generativeai
google-genai
transformers

# Frontend
streamlit
requests

# Optional
faiss-cpu
pytest
httpx
